Construct the nodes before inserting into store
Introduce abstractions for getting nodes from the store and
building up the nodes before inserting them into the store.
The reason we changed from inserting the nodes and then putting in the
parent nodes is to be consistent with the cxenstore. It ensures that all
of the nodes can be constructed first before trying to insert them into
the store, then it adds the children first, bringing the parents (and
any changes to the parents) in last.  Additionally, this
reimplementation is a bit more functional, since we are not modifying
existing nodes but are instead inserting new, modified nodes into the
store.