Refactor the data store to simplify watches
Instead of having the data store directly modify itself, the data store
returns a change set that can later be applied to the data store to
update it. Additionally, the apply process returns a set of applied
changes that can be directly processed by the watch code to generate a
list of watch events without being coupled to the data store.