message: fix the null-termination/separation issue
In general, we do not need to null-terminate strings that are passed
back to the clients. However, there is a special case where the
subdirectories that as passed back are null-terminated. So, take the
null-termination code out of the to_vec()/len() code and place it
properly in the directory code.

Additionally, we do not need to ensure that the ingress messages are
null terminated. They definitely do not have to be (such as with the
write message).