server: use a txQ instead of separate write states
Refactor the WriteHeader/WriteBody states in favor of a single Write
state. The state machine will stay in this write state until all data
is drained from the transmission queue. This allows the possibility of
enqueuing watch events into the transmission queue and having them sent
out in a future commit.